+++ ostap/fitting/tests/test_fitting_fill.py
@@ -130,6 +130,9 @@ def test_fitting_fill_1 () :
         files = prepare_data ( 4 , 5000 )
         data  = Data ( 'S' ,  files )
 
+    chain = data.chain
+
+    
     mJPsi = ROOT.RooRealVar ( 'mJPsi' , 'mass(J/Psi) [GeV]' , 3.0 * GeV , 3.2 * GeV )
 
     # =========================================================================
@@ -140,34 +143,38 @@ def test_fitting_fill_1 () :
         Variable ( mJPsi     , accessor = 'mass' ) ,
         Variable ( 'massMeV' , 'mass in MeV' , 3000 , 3200 , 'mass*1000.0' ) , 
         Variable ( 'vv102'   , 'vv10[2]'     , -1   ,  100 , '1.0*vv10[2]' ) ,
-        ( 'evt' , ) , 
+        Variable ( 'fevt'    , accessor = '1.0*evt' ) , 
         ( 'pt'  , ) ,
         ( 'eta' , ) ,
         ( 'x'   , 'some variable'  , 0 , 5000 , '(mass+pt+eta)/eta' ) 
         ]
     
-    config = { 'variables' : variables  , 'selection' : "pt>7 && eta<3"  }
+    config = { 'variables' : variables , 'selection' : "pt>7 && eta<3"  }
     
     with timing ( "No SHORTCUT, no FRAME" , logger = None ) as t1 :
         logger.info ( attention ( t1.name ) )
         selector = SelectorWithVars ( **config ) 
-        data.chain.fill_dataset ( selector , shortcut = False , use_frame = False )
-        ds1_1 = selector.data 
+        chain.fill_dataset ( selector , shortcut = False , use_frame = False )
+        ds1_1 = selector.data
     with timing ( "   SHORTCUT, no FRAME" , logger = None ) as t2 :
         logger.info ( attention ( t2.name ) )        
         selector = SelectorWithVars ( **config ) 
-        data.chain.fill_dataset ( selector , shortcut = True  , use_frame = False )        
-        ds1_2 = selector.data 
+        chain.fill_dataset ( selector , shortcut = True  , use_frame = False )        
+        ds1_2 = selector.data
     with timing ( "No SHORTCUT,    FRAME" , logger = None ) as t3 : 
         logger.info ( attention ( t3.name ) )        
         selector = SelectorWithVars ( **config ) 
-        data.chain.fill_dataset ( selector , shortcut = False , use_frame = True  )
+        chain.fill_dataset ( selector , shortcut = False , use_frame = True  )
         ds1_3 = selector.data 
     with timing ( "   SHORTCUT,    FRAME" , logger = None ) as t4 : 
         logger.info ( attention ( t4.name ) )        
         selector = SelectorWithVars ( **config ) 
-        data.chain.fill_dataset ( selector , shortcut = True  , use_frame = True  )
-        ds1_4 = selector.data 
+        chain.fill_dataset ( selector , shortcut = True  , use_frame = True  )
+        ds1_4 = selector.data        
+    with timing ( " pure-FRAME (new) " , logger = None ) as t5 : 
+        logger.info ( attention ( t5.name ) )        
+        ds1_5 , _ = chain.make_dataset ( silent = False , **config )
+
 
     table = [ ('Configuration' , 'CPU' ) ] 
 
@@ -175,30 +182,31 @@ def test_fitting_fill_1 () :
     table.append ( ( t2.name , '%.3fs' % t2.delta ) )
     table.append ( ( t3.name , '%.3fs' % t3.delta ) )
     table.append ( ( t4.name , '%.3fs' % t4.delta ) )
+    table.append ( ( t5.name , '%.3fs' % t5.delta ) )
 
     title1 = "All trivial variables"
     table1 = T.table ( table , title = title1 , prefix = '# ' , alignment = 'rr' )
     logger.info ( '%s\n%s' % ( title1 , table1 ) ) 
-    
+
     with timing ( "No SHORTCUT, no FRAME" , logger = None ) as t1 :
         logger.info ( attention ( t1.name ) )
         selector = SelectorWithVars ( **config ) 
-        data.chain.parallel_fill ( selector , shortcut = False , use_frame = False , max_files = 1 )
+        chain.parallel_fill ( selector , shortcut = False , use_frame = False , max_files = 1 )
         ds1p_1 = selector.data 
     with timing ( "   SHORTCUT, no FRAME" , logger = None ) as t2 :
         logger.info ( attention ( t2.name ) )        
         selector = SelectorWithVars ( **config ) 
-        data.chain.parallel_fill ( selector , shortcut = True  , use_frame = False , max_files = 1 )        
+        chain.parallel_fill ( selector , shortcut = True  , use_frame = False , max_files = 1 )        
         ds1p_2 = selector.data 
     with timing ( "No SHORTCUT,    FRAME" , logger = None ) as t3 : 
         logger.info ( attention ( t3.name ) )        
         selector = SelectorWithVars ( **config ) 
-        data.chain.parallel_fill ( selector , shortcut = False , use_frame = True  , max_files = 1 )
+        chain.parallel_fill ( selector , shortcut = False , use_frame = True  , max_files = 1 )
         ds1p_3 = selector.data 
     with timing ( "   SHORTCUT,    FRAME" , logger = None ) as t4 : 
         logger.info ( attention ( t4.name ) )        
         selector = SelectorWithVars ( **config ) 
-        data.chain.parallel_fill ( selector , shortcut = True  , use_frame = True  , max_files = 1 )
+        chain.parallel_fill ( selector , shortcut = True  , use_frame = True  , max_files = 1 )
         ds1p_4 = selector.data 
 
     table = [ ('Configuration' , 'CPU' ) ] 
@@ -221,7 +229,7 @@ def test_fitting_fill_1 () :
         Variable ( mJPsi     , accessor = 'mass' ) ,
         Variable ( 'massMeV' , 'mass in MeV' , 3000 , 3200 , 'mass*1000'   ) , 
         Variable ( 'vv102'   , 'vv10[2]'     , -1   ,  100 , '1.0*vv10[2]' ) , 
-        ( 'evt' , ) , 
+        Variable ( 'fevt'    , accessor = '1.0*evt' ) , 
         ( 'pt'  , ) ,
         ( 'eta' , ) ,
         ( 'x'   , 'some variable'  , 0 , 5000 , '(mass+pt+eta)/eta' ) 
@@ -243,22 +251,22 @@ def test_fitting_fill_1 () :
     with timing ( "No SHORTCUT, no FRAME" , logger = None ) as t1 : 
         logger.info ( attention ( t1.name ) )
         selector = SelectorWithVars ( **config ) 
-        data.chain.fill_dataset ( selector , shortcut = False , use_frame = False )
+        chain.fill_dataset ( selector , shortcut = False , use_frame = False )
         ds2_1 = selector.data 
     with timing ( "   SHORTCUT, no FRAME" , logger = None ) as t2 : 
         logger.info ( attention ( t2.name ) )
         selector = SelectorWithVars ( **config ) 
-        data.chain.fill_dataset ( selector , shortcut = True  , use_frame = False )
+        chain.fill_dataset ( selector , shortcut = True  , use_frame = False )
         ds2_2 = selector.data 
     with timing ( "No SHORTCUT,    FRAME" , logger = None ) as t3 : 
         logger.info ( attention ( t3.name ) )
         selector = SelectorWithVars ( **config ) 
-        data.chain.fill_dataset ( selector , shortcut = False , use_frame = True  )
+        chain.fill_dataset ( selector , shortcut = False , use_frame = True  )
         ds2_3 = selector.data 
     with timing ( "   SHORTCUT,    FRAME" , logger = None ) as t4 : 
         logger.info ( attention ( t4.name ) )
         selector = SelectorWithVars ( **config ) 
-        data.chain.fill_dataset ( selector , shortcut = True  , use_frame = True  )
+        chain.fill_dataset ( selector , shortcut = True  , use_frame = True  )
         ds2_4 = selector.data 
 
     table = [ ('Configuration' , 'CPU' ) ] 
@@ -275,22 +283,22 @@ def test_fitting_fill_1 () :
     with timing ( "No SHORTCUT, no FRAME" , logger = None ) as t1 : 
         logger.info ( attention ( t1.name ) )
         selector = SelectorWithVars ( **config ) 
-        data.chain.parallel_fill ( selector , shortcut = False , use_frame = False , maX_files = 1 )
+        chain.parallel_fill ( selector , shortcut = False , use_frame = False , maX_files = 1 )
         ds2p_1 = selector.data 
     with timing ( "   SHORTCUT, no FRAME" , logger = None ) as t2 : 
         logger.info ( attention ( t2.name ) )
         selector = SelectorWithVars ( **config ) 
-        data.chain.parallel_fill ( selector , shortcut = True  , use_frame = False , maX_files = 1 )
+        chain.parallel_fill ( selector , shortcut = True  , use_frame = False , maX_files = 1 )
         ds2p_2 = selector.data 
     with timing ( "No SHORTCUT,    FRAME" , logger = None ) as t3 : 
         logger.info ( attention ( t3.name ) )
         selector = SelectorWithVars ( **config ) 
-        data.chain.parallel_fill ( selector , shortcut = False , use_frame = True  , max_files = 1 )
+        chain.parallel_fill ( selector , shortcut = False , use_frame = True  , max_files = 1 )
         ds2p_3 = selector.data 
     with timing ( "   SHORTCUT,    FRAME" , logger = None ) as t4 : 
         logger.info ( attention ( t4.name ) )
         selector = SelectorWithVars ( **config ) 
-        data.chain.parallel_fill ( selector , shortcut = True  , use_frame = True  , max_files = 1 )
+        chain.parallel_fill ( selector , shortcut = True  , use_frame = True  , max_files = 1 )
         ds2p_4 = selector.data 
 
     table = [ ('Configuration' , 'CPU' ) ] 
@@ -314,8 +322,8 @@ def test_fitting_fill_1 () :
         variables = [
             Variable ( mJPsi     , accessor = 'mass' ) ,
             Variable ( 'massMeV' , 'mass in MeV' , 3000 , 3200  , 'mass*1000'   ) , 
-            Variable ( 'vv102'   , 'vv10[2]'     , -1   ,  100  , '1.0*vv10[2]' ) , 
-            ( 'evt' , ) , 
+            Variable ( 'vv102'   , 'vv10[2]'     , -1   ,  100  , '1.0*vv10[2]' ) ,
+            Variable ( 'fevt'    , accessor = '1.0*evt' ) , 
             ( 'pt'  , ) ,
             ( 'eta' , ) ,
             ( 'x'   , 'some variable'  , 0 , 5000 , lambda s : (s.mass+s.pt+s.eta)/s.eta ) 
@@ -328,7 +336,7 @@ def test_fitting_fill_1 () :
             Variable ( mJPsi     , accessor = 'mass' ) ,
             Variable ( 'massMeV' , 'mass in MeV' , 3000 , 3200  , 'mass*1000'   ) , 
             Variable ( 'vv102'   , 'vv10[2]'     , -1   ,  100  , '1.0*vv10[2]' ) , 
-            ( 'evt' , ) , 
+            Variable ( 'fevt'    , accessor = '1.0*evt' ) , 
             ( 'pt'  , ) ,
             ( 'eta' , ) ,
             ( 'x'   , 'some variable'  , 0 , 5000 , xvar ) 
@@ -341,22 +349,22 @@ def test_fitting_fill_1 () :
     with timing ( "No SHORTCUT, no FRAME" , logger = None ) as t1 : 
         logger.info ( attention ( t1.name ) )
         selector = SelectorWithVars ( **config ) 
-        data.chain.fill_dataset ( selector , shortcut = False , use_frame = False )
+        chain.fill_dataset ( selector , shortcut = False , use_frame = False )
         ds3_1 = selector.data 
     with timing ( "   SHORTCUT, no FRAME" , logger = None ) as t2 : 
         logger.info ( attention ( t2.name ) )
         selector = SelectorWithVars ( **config ) 
-        data.chain.fill_dataset ( selector , shortcut = True  , use_frame = False )
+        chain.fill_dataset ( selector , shortcut = True  , use_frame = False )
         ds3_2 = selector.data 
     with timing ( "No SHORTCUT,    FRAME" , logger = None ) as t3 : 
         logger.info ( attention ( t3.name ) )
         selector = SelectorWithVars ( **config ) 
-        data.chain.fill_dataset ( selector , shortcut = False , use_frame = True  )
+        chain.fill_dataset ( selector , shortcut = False , use_frame = True  )
         ds3_3 = selector.data 
     with timing ( "   SHORTCUT,    FRAME" , logger = None ) as t4 : 
         logger.info ( attention ( t4.name ) )
         selector = SelectorWithVars ( **config ) 
-        data.chain.fill_dataset ( selector , shortcut = True  , use_frame = True  )
+        chain.fill_dataset ( selector , shortcut = True  , use_frame = True  )
         ds3_4 = selector.data 
 
     table = [ ('Configuration' , 'CPU' ) ] 
@@ -374,22 +382,22 @@ def test_fitting_fill_1 () :
     with timing ( "No SHORTCUT, no FRAME" , logger = None ) as t1 : 
         logger.info ( attention ( t1.name ) )
         selector = SelectorWithVars ( **config ) 
-        data.chain.parallel_fill ( selector , shortcut = False , use_frame = False , max_files = 1 )
+        chain.parallel_fill ( selector , shortcut = False , use_frame = False , max_files = 1 )
         ds3p_1 = selector.data 
     with timing ( "   SHORTCUT, no FRAME" , logger = None ) as t2 : 
         logger.info ( attention ( t2.name ) )
         selector = SelectorWithVars ( **config ) 
-        data.chain.parallel_fill ( selector , shortcut = True  , use_frame = False , max_files = 1 )
+        chain.parallel_fill ( selector , shortcut = True  , use_frame = False , max_files = 1 )
         ds3p_2 = selector.data 
     with timing ( "No SHORTCUT,    FRAME" , logger = None ) as t3 : 
         logger.info ( attention ( t3.name ) )
         selector = SelectorWithVars ( **config ) 
-        data.chain.parallel_fill ( selector , shortcut = False , use_frame = True  , max_files = 1 )
+        chain.parallel_fill ( selector , shortcut = False , use_frame = True  , max_files = 1 )
         ds3p_3 = selector.data 
     with timing ( "   SHORTCUT,    FRAME" , logger = None ) as t4 : 
         logger.info ( attention ( t4.name ) )
         selector = SelectorWithVars ( **config ) 
-        data.chain.parallel_fill ( selector , shortcut = True  , use_frame = True  , max_files = 1)
+        chain.parallel_fill ( selector , shortcut = True  , use_frame = True  , max_files = 1)
         ds3p_4 = selector.data 
 
     table = [ ('Configuration' , 'CPU' ) ] 
@@ -414,7 +422,7 @@ def test_fitting_fill_1 () :
             Variable ( mJPsi     , accessor = 'mass' ) ,
             Variable ( 'massMeV' , 'mass in MeV' , 3000 , 3200  , 'mass*1000'   ) , 
             Variable ( 'vv102'   , 'vv10[2]'     , -1   ,  100  , '1.0*vv10[2]' ) , 
-            ( 'evt' , ) , 
+            Variable ( 'fevt'    , accessor = '1.0*evt' ) , 
             ( 'pt'  , ) ,
             ( 'eta' , ) ,
             ( 'x'   , 'some variable'  , 0 , 5000 , lambda s : (s.mass+s.pt+s.eta)/s.eta ) 
@@ -427,7 +435,7 @@ def test_fitting_fill_1 () :
             Variable ( mJPsi     , accessor = 'mass' ) ,
             Variable ( 'massMeV' , 'mass in MeV' , 3000 , 3200  , 'mass*1000'   ) , 
             Variable ( 'vv102'   , 'vv10[2]'     , -1   ,  100  , '1.0*vv10[2]' ) , 
-            ( 'evt' , ) , 
+            Variable ( 'fevt'    , accessor = '1.0*evt' ) , 
             ( 'pt'  , ) ,
             ( 'eta' , ) ,
             ( 'x'   , 'some variable'  , 0 , 5000 , xvar ) 
@@ -449,22 +457,22 @@ def test_fitting_fill_1 () :
     with timing ( "No SHORTCUT, no FRAME" , logger = None ) as t1 : 
         logger.info ( attention ( t1.name ) )
         selector = SelectorWithVars ( **config ) 
-        data.chain.fill_dataset ( selector , shortcut = False , use_frame = False )
+        chain.fill_dataset ( selector , shortcut = False , use_frame = False )
         ds4_1 = selector.data 
     with timing ( "   SHORTCUT, no FRAME" , logger = None ) as t2 : 
         logger.info ( attention ( t2.name ) )
         selector = SelectorWithVars ( **config ) 
-        data.chain.fill_dataset ( selector , shortcut = True  , use_frame = False )
+        chain.fill_dataset ( selector , shortcut = True  , use_frame = False )
         ds4_2 = selector.data 
     with timing ( "No SHORTCUT,    FRAME" , logger = None ) as t3 : 
         logger.info ( attention ( t3.name ) )
         selector = SelectorWithVars ( **config ) 
-        data.chain.fill_dataset ( selector , shortcut = False , use_frame = True  )
+        chain.fill_dataset ( selector , shortcut = False , use_frame = True  )
         ds4_3 = selector.data 
     with timing ( "   SHORTCUT,    FRAME" , logger = None ) as t4 : 
         logger.info ( attention ( t4.name ) )
         selector = SelectorWithVars ( **config ) 
-        data.chain.fill_dataset ( selector , shortcut = True  , use_frame = True  )
+        chain.fill_dataset ( selector , shortcut = True  , use_frame = True  )
         ds4_4 = selector.data 
 
     table = [ ('Configuration' , 'CPU' ) ] 
@@ -482,22 +490,22 @@ def test_fitting_fill_1 () :
     with timing ( "No SHORTCUT, no FRAME" , logger = None ) as t1 : 
         logger.info ( attention ( t1.name ) )
         selector = SelectorWithVars ( **config ) 
-        data.chain.parallel_fill ( selector , shortcut = False , use_frame = False , max_files = 1 )
+        chain.parallel_fill ( selector , shortcut = False , use_frame = False , max_files = 1 )
         ds4p_1 = selector.data 
     with timing ( "   SHORTCUT, no FRAME" , logger = None ) as t2 : 
         logger.info ( attention ( t2.name ) )
         selector = SelectorWithVars ( **config ) 
-        data.chain.parallel_fill ( selector , shortcut = True  , use_frame = False , max_files = 1 )
+        chain.parallel_fill ( selector , shortcut = True  , use_frame = False , max_files = 1 )
         ds4p_2 = selector.data 
     with timing ( "No SHORTCUT,    FRAME" , logger = None ) as t3 : 
         logger.info ( attention ( t3.name ) )
         selector = SelectorWithVars ( **config ) 
-        data.chain.parallel_fill ( selector , shortcut = False , use_frame = True  , max_files = 1 )
+        chain.parallel_fill ( selector , shortcut = False , use_frame = True  , max_files = 1 )
         ds4p_3 = selector.data 
     with timing ( "   SHORTCUT,    FRAME" , logger = None ) as t4 : 
         logger.info ( attention ( t4.name ) )
         selector = SelectorWithVars ( **config ) 
-        data.chain.parallel_fill ( selector , shortcut = True  , use_frame = True  , max_files = 1 )
+        chain.parallel_fill ( selector , shortcut = True  , use_frame = True  , max_files = 1 )
         ds4p_4 = selector.data 
 
     table = [ ('Configuration' , 'CPU' ) ] 
@@ -515,6 +523,7 @@ def test_fitting_fill_1 () :
     if ds1_1 != ds1_2   : logger.warning  ('Datasets ds1_1  and ds1_2   are different!' )
     if ds1_1 != ds1_3   : logger.warning  ('Datasets ds1_1  and ds1_3   are different!' )
     if ds1_1 != ds1_4   : logger.warning  ('Datasets ds1_1  and ds1_4   are different!' )
+    if ds1_1 != ds1_5   : logger.warning  ('Datasets ds1_1  and ds1_5   are different!' )
 
     if ds1_1 != ds1p_1  : logger.warning  ('Datasets ds1_1  and ds1p_1  are different!' )
     if ds1_2 != ds1p_2  : logger.warning  ('Datasets ds1_2  and ds1p_2  are different!' )
@@ -565,12 +574,12 @@ def test_fitting_fill_1 () :
     logger.info ( '%s\n%s' % ( title4  , table4  ) ) 
     logger.info ( '%s\n%s' % ( title4p , table4p ) ) 
 
-    
 # =============================================================================
 if '__main__' == __name__ :
 
-    
+
     test_fitting_fill_1 ()
+
 
 # =============================================================================
 ##                                                                      The END